requirements.txt
terminaltables==3.1.10
requests==2.29.0
python-dotenv==1.0.0